+++ Project.toml
@@ -6,9 +6,16 @@ version = "0.2.0"
 [deps]
 Dictionaries = "85a47980-9c8c-11e8-2b9f-f7ca1fa99fb4"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
 NamedGraphs = "678767b0-92e7-4007-89e4-4527a8725b19"
 SimpleTraits = "699a6c99-e7fa-54fc-8d76-47d257e15c1d"
+
+[weakdeps]
+GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
+# NamedGraphs = "678767b0-92e7-4007-89e4-4527a8725b19"
+
+[extensions]
+DataGraphsGraphsFlowsExt = "GraphsFlows"
+# DataGraphsNamedGraphsExt = "GraphsFlows"
 
 [compat]
 Dictionaries = "0.4"
@@ -19,6 +26,8 @@ SimpleTraits = "0.9"
 julia = "1.7"
 
 [extras]
+GraphsFlows = "06909019-6f44-4949-96fc-b9d9aaa02889"
+# NamedGraphs = "678767b0-92e7-4007-89e4-4527a8725b19"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
