+++ src/traits/isunderlyinggraph.jl
@@ -10,6 +10,3 @@ is_underlying_graph(::Type{<:AbstractGraph}) = false
 
 using Graphs.SimpleGraphs: AbstractSimpleGraph
 is_underlying_graph(::Type{<:AbstractSimpleGraph}) = true
-
-using NamedGraphs: AbstractNamedGraph
-is_underlying_graph(::Type{<:AbstractNamedGraph}) = true
