+++ test/runtests.jl
@@ -17,9 +17,11 @@ using Graphs:
   ne,
   nv,
   outdegree,
+  path_graph,
   src,
   vertices
 using Graphs.SimpleGraphs: SimpleDiGraph, SimpleEdge, SimpleGraph
+using GraphsFlows: GraphsFlows
 using NamedGraphs: NamedDiGraph, NamedGraph
 using NamedGraphs.GraphsExtensions: ⊔, rename_vertices
 using NamedGraphs.NamedGraphGenerators: named_grid, named_path_graph
@@ -338,6 +340,16 @@ using DataGraphs: is_arranged
     @test ps.dists == dictionary([1 => 0, 2 => 1, 3 => 2, 4 => 3])
     @test ps.parents == dictionary([1 => 1, 2 => 1, 3 => 2, 4 => 3])
     @test ps.pathcounts == dictionary([1 => 1.0, 2 => 1.0, 3 => 1.0, 4 => 1.0])
+  end
+  @testset "GraphsFlows.mincut (vertextype=$(eltype(verts))" for verts in (
+      [1, 2, 3, 4],
+      ["A", "B", "C", "D"],
+    )
+      g = DataGraph(NamedGraph(path_graph(4), verts))
+      part1, part2, flow = GraphsFlows.mincut(g, verts[1], verts[4])
+      @test verts[1] ∈ part1
+      @test verts[4] ∈ part2
+      @test flow == 1
   end
 end
 end
